Further warning elimination

+++ SolastaCommunityExpansion/Subclasses/Barbarian/PathOfTheLight.cs
@@ -185,9 +185,7 @@ namespace SolastaCommunityExpansion.Subclasses.Barbarian
 
         private static void ApplyLightsProtectionHealing(ulong sourceGuid)
         {
-            var conditionSource = RulesetEntity.GetEntity<RulesetCharacter>(sourceGuid) as RulesetCharacterHero;
-
-            if (conditionSource == null || conditionSource.IsDead)
+            if (!(RulesetEntity.GetEntity<RulesetCharacter>(sourceGuid) is RulesetCharacterHero conditionSource) || conditionSource.IsDead)
             {
                 return;
             }
@@ -485,24 +483,19 @@ namespace SolastaCommunityExpansion.Subclasses.Barbarian
 
         private static void HandleAfterIlluminatedConditionRemoved(RulesetActor removedFrom)
         {
-            var character = removedFrom as RulesetCharacter;
-
-            if (character == null)
+            if (!(removedFrom is RulesetCharacter character))
             {
                 return;
             }
 
             // Intentionally *includes* conditions that have Illuminated as their parent (like the Illuminating Burst condition)
-            if (!character.HasConditionOfTypeOrSubType(IlluminatedConditionName))
+            if (!character.HasConditionOfTypeOrSubType(IlluminatedConditionName)
+                && (character.PersonalLightSource?.SourceName == IlluminatingStrikeName || character.PersonalLightSource?.SourceName == IlluminatingBurstName))
             {
-                if (character.PersonalLightSource?.SourceName == IlluminatingStrikeName ||
-                    character.PersonalLightSource?.SourceName == IlluminatingBurstName)
-                {
-                    var visibilityService = ServiceRepository.GetService<IGameLocationVisibilityService>();
+                var visibilityService = ServiceRepository.GetService<IGameLocationVisibilityService>();
 
-                    visibilityService.RemoveCharacterLightSource(GameLocationCharacter.GetFromActor(removedFrom), character.PersonalLightSource);
-                    character.PersonalLightSource = null;
-                }
+                visibilityService.RemoveCharacterLightSource(GameLocationCharacter.GetFromActor(removedFrom), character.PersonalLightSource);
+                character.PersonalLightSource = null;
             }
         }
